app/requirements-minimal.txt
# Minimal requirements for initial API startup
# Core web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==22.0.0

# Database and models
sqlmodel==0.0.22
pydantic[email]==2.9.2
pydantic-settings==2.5.2

# Azure essentials
azure-storage-blob==12.23.1
azure-identity==1.18.0
azure-cosmos==4.7.0
azure-core==1.31.0

# Service Bus for messaging
azure-servicebus==7.12.3

# Basic utilities
python-dotenv==1.0.1
python-dateutil==2.9.0
aiofiles==24.1.0
aiohttp==3.10.10
cachetools==5.5.0

# Logging
psutil==6.0.0